Start Client Creation

+++ src/data_processing.py
@@ -181,7 +181,7 @@ def create_dataloader(X, y):
     class SisFallDataset(Dataset):
         def __init__(self, X, y):
             self.X = torch.from_numpy(X).float()
-            self.y = torch.from_numpy(y).float()
+            self.y = torch.from_numpy(y).int()
         def __len__(self): return len(self.X)
         def __getitem__(self, idx): 
             # Transpose [Time, Channel] -> [Channel, Time] for CNN
